assign sympy version when py>=3.9

+++ onnxruntime/python/tools/symbolic_shape_infer.py
@@ -1679,7 +1679,7 @@ class SymbolicShapeInference:
         if get_opset(self.out_mp_) <= 10:
             scales = self._try_get_value(node, 1)
             if scales is not None:
-                new_sympy_shape = [sympy.simplify(float(sympy.floor(d * s))) for d, s in zip(input_sympy_shape, scales)]
+                new_sympy_shape = [sympy.simplify(sympy.floor(d * s)) for d, s in zip(input_sympy_shape, scales)]
                 self._update_computed_dims(new_sympy_shape)
                 vi.CopyFrom(
                     helper.make_tensor_value_info(
@@ -1693,7 +1693,7 @@ class SymbolicShapeInference:
             scales = self._try_get_value(node, 2)
             sizes = self._try_get_value(node, 3)
             if sizes is not None:
-                new_sympy_shape = [sympy.simplify(float(sympy.floor(s))) for s in sizes]
+                new_sympy_shape = [sympy.simplify(sympy.floor(s)) for s in sizes]
                 self._update_computed_dims(new_sympy_shape)
             elif scales is not None:
                 rank = len(scales)
@@ -1706,7 +1706,7 @@ class SymbolicShapeInference:
                     roi_end = [1] * rank
                 scales = list(scales)
                 new_sympy_shape = [
-                    sympy.simplify(float(sympy.floor(d * (end - start) * scale)))
+                    sympy.simplify(sympy.floor(d * (end - start) * scale))
                     for d, start, end, scale in zip(input_sympy_shape, roi_start, roi_end, scales)
                 ]
                 self._update_computed_dims(new_sympy_shape)
